+++ doc/make.jl
@@ -195,7 +195,6 @@ BaseDocs = [
 StdlibDocs = [stdlib.targetfile for stdlib in STDLIB_DOCS]
 
 Tutorials = [
-    "tutorials/creating-packages.md",
     "tutorials/profile.md",
     "tutorials/external.md",
 ]